+++ src/apps/notifications/components/NotificationDetails.tsx
@@ -21,6 +21,10 @@ import {
   NotificationItem
 } from "../types";
 import { EnhancedNotificationSummaryTableHead } from "./EnhancedNotificationSummaryTableHead"
+import {
+  getComparator,
+  stableSort
+} from "../../../utils";
 
 interface Props {
   open: boolean;
@@ -68,11 +72,7 @@ const NotificationDetails: React.FC<Props> = ({ open, handleClose, notification 
   const getConceptType = (row: NotificationItemRow) => {
     const getActualIndex = (ele: string) => ele ===  getConceptId(row);
     const actualIndex = parentConceptIds.findIndex(getActualIndex);
-    if (actualIndex > -1) {
-      return "Parent";
-    } else {
-      return "Dependent";
-    }
+    return actualIndex > -1 ? "Parent" : "Dependent";
   };
 
   const getStatus = (row: NotificationItemRow) => {
@@ -80,10 +80,7 @@ const NotificationDetails: React.FC<Props> = ({ open, handleClose, notification 
   };
 
   const getReason = (row: NotificationItemRow) => {
-    if (typeof row.message === "string") {
-      return row.message;
-    }
-    return row.message.join(" ");
+    return (typeof row.message === "string") ? row.message : row.message.join(" ");
   };
 
   const getDialogTitle = () => {
@@ -127,35 +124,6 @@ const NotificationDetails: React.FC<Props> = ({ open, handleClose, notification 
     setOrderBy(property);
     setPage(0);
   };
-
-  function descendingComparator<T>(a: T, b: T, orderBy: keyof T) {
-    if (b[orderBy] < a[orderBy]) {
-      return -1;
-    }
-    if (b[orderBy] > a[orderBy]) {
-      return 1;
-    }
-    return 0;
-  }
-
-  function getComparator<Key extends keyof any>(
-      order: Order,
-      orderBy: Key,
-  ): (a: { [key in Key]: number | string }, b: { [key in Key]: number | string }) => number {
-    return order === 'desc'
-        ? (a, b) => descendingComparator(a, b, orderBy)
-        : (a, b) => -descendingComparator(a, b, orderBy);
-  }
-
-  function stableSort<T>(array: T[], comparator: (a: T, b: T) => number) {
-    const stabilizedThis = array.map((el, index) => [el, index] as [T, number]);
-    stabilizedThis.sort((a, b) => {
-      const order = comparator(a[0], b[0]);
-      if (order !== 0) return order;
-      return a[1] - b[1];
-    });
-    return stabilizedThis.map((el) => el[0]);
-  }
 
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(10);
